Refactor DTA solvers: extract C++ engine code into DTAsolvers_cpp_adapter
Move CppSolverDUE/CppSolverDSO_D2D and their helpers (id<->name
conversion, CSR route handling, lightweight finalize) from DTAsolvers.py
into a dedicated DTAsolvers_cpp_adapter.py module (pure code move, no
logic changes). The core module now references the C++ path only via a
lazy import inside SolverDUE/SolverDSO_D2D.__new__ plus an
_is_cpp_solver init guard, so the adapter (and the C++ extension) is
never imported in pure-Python usage.

Bit-identical DUE/DSO results. 200 cpp-mode tests and 4 Python-mode DTA
verification tests pass.
